Correcciones y repositorio de ingredientesReceta

+++ Gourmet/Repositories/Contracts/IRecetaRepository.cs
@@ -4,7 +4,7 @@ namespace Gourmet.Repositories.Contracts
     public interface IRecetaRepository
     {
         IEnumerable<Receta> GetRecetas();
-        Receta GetRecetaByID(int recetaId);
+        Receta GetRecetaByID(long recetaId);
         void InsertReceta(Receta receta);
         void DeleteReceta(int recetaID);
         void UpdateReceta(Receta receta);
